requirements.txt
yfinance
google-generativeai
requests
edge-tts
moviepy
Pillow
google-api-python-client
google-auth-oauthlib
google-auth-httplib2
rich
python-dotenv
